subprocess: Add and use _srt_subprocess_runner_spawn_sync()
This encapsulates `g_spawn_sync()` with a useful set of flags and
behaviours, and will let us move the responsibility for handling
features like timeouts between `get_helper()` and `spawn_sync()`.
To make that work, try harder to pass the same flags to both
`get_helper()` and `spawn_sync()`.

For `architecture`, `display`, `graphics-drivers` and `libdl`, no
behaviour change is intended.

`display` and `graphics-drivers` should probably not be using
`KEEP_GAMEOVERLAYRENDERER` (there's no real reason to do so) but for
now, preserve existing behaviour.

For `library`, we were previously passing `G_SPAWN_SEARCH_PATH` to
`g_spawn_sync()`, but it would have had no practical effect, because
we're using the absolute path to a helper executable. Drop that flag.

For `graphics`, we were previously searching the `PATH` for everything
except `check-gl`. Consistently search the `PATH` in all cases.

Signed-off-by: Simon McVittie <smcv@collabora.com>