pv-wrap: Don't convert pv_wrap_share_sockets() paths to interpreter root
Most of what we export via FlatpakExports or pv_wrap_share_sockets either
is meant to show up at a corresponding path inside the container's real
root, for example /home -> /home, or belongs to /var and is meant to
show up at a corresponding path in the real root and/or the $FEX_ROOTFS,
for example /var/tmp -> /var/tmp.

However, some paths that come via pv_wrap_share_sockets() are
bind-mounts with their destination inside /run/host, for example
$FEX_ROOTFS/usr/share/fonts -> /run/host/fonts. We want to mount those
in the real root, not in the $FEX_ROOTFS, to avoid a split-brain
situation where ARM and x86 code sees a different /run.

Signed-off-by: Simon McVittie <smcv@collabora.com>