pv-wrap: Move environment variable setup into bwrap.c
wrap.c is too big, and because it contains main(), it's hard to unit-test.

No functional changes intended, but the structure of this code has been
changed: the difference between the "Flatpak subsandbox" and "Flatpak
with the equivalent of flatpak-spawn --host" code paths was previously
inside the loop, but is now outside the loop.

Signed-off-by: Simon McVittie <smcv@collabora.com>