subprocess: Factor out an object for the execution environment
This will let us bundle together more of the subprocess handling for
running helpers. For now, it is just a container for the environment,
helpers path and test flags.

This is an immutable value-object which does not change after it has
been constructed, meaning that we can share it between threads if
that becomes useful in future.

No functional change intended.

Signed-off-by: Simon McVittie <smcv@collabora.com>