subprocess: Avoid using timeout(1) to impose a timeout
This is more difficult than you might think, because until GLib 2.78
(which is far too new for us to depend on), there was a potential
race condition between g_child_watch_source_new()'s use of waitpid(),
and a kill() in the same main-context. We don't really want to use
g_child_watch_source_new() anyway, because we don't know whether this
library is going to be loaded by a process that is blocking SIGCHLD or
handling SIGCHLD itself.

Avoiding timeout(1) can result in a significant speedup on ARM hardware
emulating x86 via FEX-Emu, which makes executable startup
disproportionately expensive: each new executable needs to re-initialize
the interpreter, which is normally fine, but the cost builds up when we
run a very large number of very small processes.

Avoiding timeout(1) is also good for robustness: in the past, Canonical's
Steam Snap app has had an AppArmor profile that did not allow arbitrary
coreutils commands to be run.

In this initial implementation, we poll the process every 100ms. This
is *much* slower than blocking (it adds up to 100ms to every subprocess
execution, resulting in a slowdown from around 8 to 96 seconds for
steam-runtime-system-info on x86), but it can be used as a fallback.
It's in this commit to have an opportunity to test the fallback code path.
Subsequent commits will speed this up again.

Helps: steam-runtime-tools#88
Signed-off-by: Simon McVittie <smcv@collabora.com>