subprocess: Try to wait for subprocess in a thread instead of polling
The key trick here is that on any halfway modern kernel, we can use
waitid() in a background thread with the WNOWAIT flag, which pauses until
the subprocess has exited but does not reap it, allowing the main thread
to wait for it again, this time without WNOWAIT, causing the subprocess
to be reaped.

This is considerably faster than the previous implementation, around 30%
on both x86 and ARM.

I considered using a GThreadPool to reuse the short-lived threads,
but that's more complex and doesn't actually seem to be any faster
than creating and destroying the thread every time.

Resolves: steam-runtime-tools#88
Signed-off-by: Simon McVittie <smcv@collabora.com>