check-xdg-portal: Allocate and free one GVariant per loop iteration
Previously we allocated one GVariant per loop, but only freed the last.
When building with AddressSanitizer, this caused a crash which is reported
to systemd-coredump if used (but the JSON output was valid, and the test
suite can't assert that check-xdg-portal succeeds because on some test
systems it might not, so the crash did not cause a test failure).

Signed-off-by: Simon McVittie <smcv@collabora.com>