pv-runtime: Go back to using /run/pressure-vessel/ldso under Snap
In v0.20231208.0 we switched this to /var/pressure-vessel/ldso to
avoid some strange behaviour under FEX-Emu. Unfortunately, snapd's
AppArmor profiles make assumptions about internal implementation details
of pressure-vessel, and do not currently allow /var/pressure-vessel
(canonical/steam-snap#356, steam-runtime#640); and none of the maintainers
or users of steam-snap seem to be using our betas, so this was only
discovered when v0.20231208.0 was promoted from beta to stable.

Fixing this requires snapd changes, which take a while to deploy, so
temporarily work around it by going back to the previous paths when
snapd is detected.

Unfortunately, this also requires us to relax some checks that were put
in place to check that we aren't going to regress FEX-Emu.

Helps: https://github.com/canonical/steam-snap/issues/356
Helps: https://github.com/ValveSoftware/steam-runtime/issues/640
Signed-off-by: Simon McVittie <smcv@collabora.com>