pv-runtime: Make dependency libraries in /var/lib/snapd/hostfs available
This allows us to make /usr/share/nvidia a symlink to
/var/lib/snapd/hostfs/usr/share/nvidia. The Nvidia drivers are currently
hard-coded to read data from /usr/share/nvidia, and I have been told
that this is functionally necessary in some situations.

As with the previous commit, before this can be enabled by default,
snapd's AppArmor profile needs to be fixed to allow it.

Helps: https://github.com/ValveSoftware/steam-runtime/issues/586
Signed-off-by: Simon McVittie <smcv@collabora.com>