SteamLinuxRuntime_sniper.sh: Tolerate CRLF line endings in VERSIONS.txt
It seems that the process of packing the runtime files into
runtime_sniper_ubuntu12.zip and unpacking them on the end-user system
translates (some?) text files' line endings from Unix LF to Windows CRLF,
which means that a simple byte-based comparison will fail.

Until we can find a better way to handle this, strip the CRs from the
line endings and compare the resulting contents as text.

Signed-off-by: Simon McVittie <smcv@collabora.com>