container-runtime: Suppress non-essential mtree manifest in many-builds.py
This is very slow if we're unpacking several sysroots into the same
top-level directory, as build-aux/many-builds.py does.

We still need to generate mtrees for any runtime that we have minimized,
because in a minimized runtime, various categories of files (zero-byte
regular files, empty directories, symlinks) only exist as records in
the mtree manifest.

Signed-off-by: Simon McVittie <smcv@collabora.com>