CI: Simplify path to relocatable amd64+i386 build of pressure-vessel
Avoid the word "production", which could easily be mistaken for an
official release. It's here for historical reasons. We used to do two
separate relocatable installations, one with production-like compiler
options, and one unoptimized and unstripped for debugging; but I removed
the latter in commit 8649f766 "CI: Skip unpackaged scout builds" in mid
2021, because it slowed down CI and was rarely useful.

Signed-off-by: Simon McVittie <smcv@collabora.com>