Rename PvBwrapLock to SrtFileLock and move it to common code
This was implemented to interact correctly with the locking scheme
used in bwrap(1) (which uses the non-OFD flavour of fnctl locks), but
is otherwise unrelated to bwrap(1) or pressure-vessel. Moving this into
common code will facilitate using it outside pressure-vessel, in
particular when unpacking `ubuntu12_64/steam-runtime-sniper/`.

While I'm changing the interface anyway, take the opportunity to rename
the WRITE flag to EXCLUSIVE, which is clearer.

Helps: steamrt/tasks#387
Signed-off-by: Simon McVittie <smcv@collabora.com>