scout-on-soldier: Remove workaround for mid 2021 configuration change
This script was added in June 2021 (T28437) when we migrated app ID
1070560 from being a self-contained "pure scout" runtime
(for which the entry point was ./_v2-entry-point) to being a layered
scout-on-soldier runtime (for which the entry point is
./scout-on-soldier-entry-point-v2). At the time, the Steam client did
not reliably pick up the configuration change until it was restarted.

It seems safe to assume that all users have restarted Steam since 2021.

Removing this workaround also incidentally removes one of our uses of
zenity, which helps steamrt/tasks#397 and related tasks.

Signed-off-by: Simon McVittie <smcv@collabora.com>