runtime: Only unset STEAM_ZENITY if it belongs to the LDLP runtime
There are currently four possibilities for STEAM_ZENITY:

* unset
* explicitly set to empty (to disable its use, as on Steam Deck)
* absolute path of host zenity, typically /usr/bin/zenity
* literal string "zenity" which will be found in scout's PATH

In all cases except the last, there is no need to unset STEAM_ZENITY
when "escaping" from the LDLP runtime, because the value will be
equally valid in a host-like execution environment; and when the value
is empty, it can be actively harmful to unset it, since this will result
in using zenity when not intended.

For completeness, we also unset STEAM_ZENITY if it is the absolute
path of the LDLP runtime's zenity (which is not a situation that Steam
ever sets up right now, but in principle it could be).

steamrt/tasks#397

Signed-off-by: Simon McVittie <smcv@collabora.com>