pv-adverb: Don't require target of --assign-fd to be a valid fd yet
If we're doing `pv-adverb --assign-fd=9=1` (the equivalent of `9>&1`
in the shell), then fd 1 needs to be a valid fd already, but fd 9
does not.

This was not a problem in practice because pressure-vessel currently
only assigns to fds 1 and 2 (stdout and stderr), which we ensure are
always valid, even if that means opening /dev/null.

Fixes: 4329cec7 "pv-adverb: Add --assign-fd"
Signed-off-by: Simon McVittie <smcv@collabora.com>