build: Only build pressure-vessel during development and for scout
The versions of pressure-vessel in official Steam Linux Runtime releases
come from builds done in the scout environment. Binary builds of
pressure-vessel in any newer environment are not very useful, because
they would require a modern version of glibc and other supporting
libraries, defeating its purpose.

If steam-runtime-tools gets packaged in Linux distributions as a source
of diagnostic tools or similar, we don't want it to include a build of
pressure-vessel that would be of little use to anyone, and mainly just
confusing.

We do still want to build pressure-vessel when developers are working
on steam-runtime-tools, though - that way, we get better test coverage
and compiler warnings. Explicitly enable it in many-builds.py.

Signed-off-by: Simon McVittie <smcv@collabora.com>