Rename pv-bwrap to srt-bwrap, and build it as part of the bin option
Now that SLR - sniper is a mandatory dependency for steamwebhelper,
we want to be able to test for a working bwrap as part of
steam-runtime-check-requirements.

Because this change alters the default_options for a subproject, it
requires build directories to be reconfigured with `meson setup --wipe`.
For users of `build-aux/many-builds.py` it should be sufficient to run:

    build-aux/many-builds.py setup

Signed-off-by: Simon McVittie <smcv@collabora.com>