pv-wrap: Explicitly share ~/.config, etc. if sharing $HOME
According to recent steamwebhelper issue reports, it's somewhat common
for ~/.config to be a symbolic link to some other filesystem, in which
case users will expect the target of the symlink to make it into the
container.

This commit only handles the case where an XDG environment variable is
unset, resulting in us using its default. In fact we share the default
directories unconditionally, because it's somewhat common for
applications to ignore the environment variable and hard-code its
default - so if we have XDG_CONFIG_HOME=$HOME/Config, it's safest for
us to assume that both ~/Config and ~/.config are potentially needed.

A separate commit will handle the case where the variable is set, which
needs to happen in a different part of the codebase.

steamrt/tasks#411

Signed-off-by: Simon McVittie <smcv@collabora.com>