pv-wrap: Stop using global state in pv_bind_and_propagate_from_environ()
This will allow us to unit-test it.

There is a very slight behaviour change here when running under FEX-Emu:
previously we would test the existence of directories from the
environment with g_file_test() (which looks in FEX-Emu's emulated root
filesystem), but now we test for their existence in /proc/self/root.
In practice we do not expect this to matter for any of the environment
variables that are mentioned in pv_bind_and_propagate_from_environ().

Signed-off-by: Simon McVittie <smcv@collabora.com>