pv-wrap: Explicitly share $XDG_CONFIG_HOME, etc. if sharing $HOME
According to recent steamwebhelper issue reports, it's somewhat common
for ~/.config to be a symbolic link to some other filesystem, in which
case users will expect the target of the symlink to make it into the
container.

This commit only handles the case where an XDG environment variable is
explicitly set. The case where it is unset (defaulting to ~/.config,
etc.) was already handled in a previous commit.

It is normal for $XDG_DATA_DIRS to contain /usr/local/share:/usr/share,
and $XDG_CONFIG_DIRS might legitimately contain /usr paths too, so
downgrade the usual warning about paths below /usr to a debug message
for these particular variables.

steamrt/tasks#411

Signed-off-by: Simon McVittie <smcv@collabora.com>