container: Do a functional test on flatpak-portal subsandboxes
This gives us some ability to diagnose what, if anything, is wrong with
them. Obviously if the Steam UI doesn't start, we will not be able to
put this information in Help -> Steam Runtime Diagnostics, but we can
call into this from steam-runtime-check-requirements, or direct users
to run steam-runtime-system-info via `flatpak run --command=bash`.

steamrt/tasks#406

Signed-off-by: Simon McVittie <smcv@collabora.com>