pressure-vessel: Generalize sharing other app frameworks' directories
In addition to Guix and NixOS, we would like to share equivalent paths
used by Snap. That gives us enough of these to justify doing this with
a loop across an array, to make it easier to add similar things for
other app frameworks later.

Unfortunately snapd's AppArmor profile doesn't yet allow sharing
the paths that we want to share in this way, so set them up as data
structures that contain workarounds flags.

Helps: https://github.com/ValveSoftware/steam-runtime/issues/586
Signed-off-by: Simon McVittie <smcv@collabora.com>