SrtEnvOverlay: Rename from PvEnviron
This isn't really an environment block, so it was always mis-named.
It's really a set of instructions for how to modify an environment block
by setting variables, unsetting variables, or allowing variables to be
inherited.

Moving it into library code will allow us to share it between
srt-launch-client and anywhere else that needs it.

Signed-off-by: Simon McVittie <smcv@collabora.com>