mtree: Don't fail verification for parents of a contents= filename
The files listed in the manifest as `contents=` are part of the files
that are physically distributed, but they are not part of the logical
hierarchy described by the manifest, so it wouldn't make sense to list
their parents as part of the manifest. This will matter if we want to
split files into subdirectories by a prefix of their hash, to avoid
having a very large number of content-addressed files at top level.

steamrt/tasks#422

Signed-off-by: Simon McVittie <smcv@collabora.com>
(cherry picked from commit d7396500de711e6842d84343f4199a717660314e)