file-lock: Add a verbose mode
If we're waiting for a lock before doing something (for example
unpacking the runtime for steamwebhelper) and we expect that under
normal circumstances the lock will not be held, it's helpful to log
a message in the rare case where we have to wait for it, so that a
reader can think "why is this locked? should it be locked?" and take
any action that might be required.

This is implemented by trying to take the lock without waiting, and
if that fails, logging a message and then retrying with the wait.

The messages will look like:

    steam-runtime-supervisor[12345]: N: Waiting for exclusive lock to be available: /path/to/lock
    ... time passes ...
    steam-runtime-supervisor[12345]: N: Acquired lock /path/to/lock, continuing

Signed-off-by: Simon McVittie <smcv@collabora.com>