SLR: Add filter_exclusive property to tool manifests
For "pure" container runtimes (soldier, sniper and future branches)
we set a low priority, so that these container runtimes will be hidden
from the Steam UI, unless forced via an intentionally undocumented
developer option. For simplicity, if scout is built as a pure container
runtime, it gets the same treatment (but in practice we do not generate
or publish a pure scout runtime any more, because "layered"
scout-on-soldier turned out to have significantly better compatibility
with real-world games).

For "layered" container runtimes (scout-on-soldier, and possibly
scout-on-sniper in future), we set a higher priority, so that these
container runtimes will be shown for legacy scout Linux games.

For games that target a later ABI, such as CS2 targeting sniper, the
Steam client is expected to overwrite the priority value set here with
a priority greater than 10, so that the appropriate runtime will be the
only one that is offered.

steamrt/tasks#426

Signed-off-by: Simon McVittie <smcv@collabora.com>