pv-wrap: Share code to handle suppressing exports that won't work
Make this a method on the PvWrapContext, so that it can use a hash table
(lazily allocated) to store the set of paths for which we already logged
a warning.

No functional change intended, except for some slightly different
warning messages.

steamrt/tasks#434

Helps: https://github.com/ValveSoftware/steam-for-linux/issues/10789
Signed-off-by: Simon McVittie <smcv@collabora.com>