FlatpakExports: Modify to use pressure-vessel's list of reserved paths
pressure-vessel reserves more paths than Flatpak does, and we need to
avoid all of them. For most of the code paths that set up exports,
we will avoid the paths anyway (in order to log warnings), but as a
fallback, it would be good to be sure that every possible reason for
paths to be exported is covered.

steamrt/tasks#434

Helps: https://github.com/ValveSoftware/steam-for-linux/issues/10789
Signed-off-by: Simon McVittie <smcv@collabora.com>