wrap-context: Treat more paths as "part of the OS"
ClearLinux adds /var/cache to the XDG_DATA_DIRS, and in a subsequent
commit we will want to ignore that with only an informational message,
rather than logging warnings about a situation that is considered correct.

Similarly, we don't want to log warnings about /etc/xdg being in
XDG_CONFIG_DIRS, and an OS might conceivably put generated files in /run
on the same basis that ClearLinux uses /var/cache.

I haven't added /bin, /lib* or /sbin here, on the basis that they are
unlikely to appear in these search paths in practice.

steamrt/tasks#434

Helps: https://github.com/ValveSoftware/steam-for-linux/issues/10789
Signed-off-by: Simon McVittie <smcv@collabora.com>