utils test: Locate test data relative to executable correctly
Because libcapsule uses non-recursive make, G_TEST_SRCDIR and
G_TEST_BUILDDIR are set to the top-level directory of the source/build
tree. When running the test suite "as-installed", the equivalent for
both is `${libexecdir}/installed-tests/libcapsule`.

This matters because we locate the "version 1" and "version 2" libraries
during testing as `${builddir}/tests/version{1,2}`, which is incorrect
(a duplicate `tests/`) if the `${builddir}` is set to the `tests`
subdirectory rather than the top-level directory.

Signed-off-by: Simon McVittie <smcv@collabora.com>