debian: Require at least automake 1.15
Steam Runtime 1 'scout' only has 1.11 as its default version, and that's
too old for AM_TESTS_ENVIRONMENT to be respected. It does have 1.15
available, so make sure we have at least that version one way or another.

At the moment this is mitigated by the fact that scout only has GLib 2.32,
which is too old, and we use --with-glib=auto to skip building and running
the one test that uses GLib (utils.t) when building on scout - which
happens to be the only test where correct handling of AM_TESTS_ENVIRONMENT
was strictly required. However, if we backport a newer GLib into scout,
we'll start running that test, and therefore we'll need it to work.

steamrt/tasks#112

Signed-off-by: Simon McVittie <smcv@collabora.com>