logger: Copy log messages to terminal/stderr without delay
If the process that is wrapped by the logger is an interactive CLI
such as apt-get, then we want prompts and similar interactive content
to get to the terminal as quickly as possible, without waiting for the
end of a line to be reached. Separate the processing of partial lines
(which we send to the terminal ASAP) from the processing of complete
lines (which we still buffer, so that we can send them to the log file
or systemd Journal as a complete line, to avoid log rotation happening
halfway through a line).

We have a fallback code path that logs to standard error, which is used
if all other log destinations (log file, terminal, Journal) are disabled.
We don't know the file type of standard error (it could be an interactive
terminal, a pipe or socket, or something else) so err on the side of
sending partial messages on that code path without delay as well.

steamrt/tasks#447

Signed-off-by: Simon McVittie <smcv@collabora.com>