logger: Add --background option
Making the long-running logging process daemonize is necessary to avoid
some failure modes when combined with steam-runtime-supervisor.

Without this option, if we run a process inside a supervisor that will
passively wait for child processes, and we want a srt-logger to log the
output of the supervisor and all of the supervised processes:

    srt-logger -d ~/tmp -t supervised-shell -- \
    steam-runtime-supervisor --subreaper -- \
    sh -c 'echo "placeholder process $$ exiting"'

then we have a deadlock. The logging process that was started internally
by srt-logger becomes a child of the steam-runtime-supervisor, which
means the supervisor will not exit until after the logger does.
Meanwhile, the logger is receiving the supervisor's stdout/stderr via a
pipe, so it will not exit until after the supervisor does, a cyclic
dependency that results in neither one exiting.

We can break the deadlock by using the new --background option:

    srt-logger --background -d ~/tmp -t supervised-shell -- \
    steam-runtime-supervisor --subreaper -- \
    sh -c 'echo "placeholder process $$ exiting"'

which makes the internal logging process daemonize (double-fork),
reparenting itself to a "larger" subreaper if present, or to process 1;
either way, the supervisor no longer needs to wait for it.

Similarly, if we set up the same scenario, but this time we want the
supervisor to terminate any leaked background processes (such as the
sleep(1) process in this example) after the main process exits:

    srt-logger -d ~/tmp -t supervised-shell -- \
    steam-runtime-supervisor --subreaper --terminate-timeout=5 -- \
    sh -c 'sleep 3600 & echo "placeholder process $$ exiting"'

then we have broken the deadlock, but at the cost of introducing a loss
of log information: the supervisor will send SIGTERM to the logger,
which will stop recording log messages, and potentially important log
messages from the supervisor or the background processes will not be
written to logs or to the terminal.

Again, we can resolve this problem by adding --background to the srt-logger
invocation, which takes it outside the supervisor's control.

The complexity cost of putting the srt-logger into the background is not
required in all cases, so --background is an option and not the default.

In all examples shown here, the sh(1) process is a placeholder
representing an arbitrary "payload": more realistically, it would be a
game, or some other component such as steamwebhelper.

steamrt/tasks#460

Signed-off-by: Simon McVittie <smcv@collabora.com>