logger: Parse log line prefixes and use them for target filtering
This adds the ability to parse syslog prefixes from log lines by passing
--parse-level-prefix to srt-logger. The corresponding lines will then be
filtered based on the newly added --[journal/file/terminal]-level
options.

The presence of prefixes means that, when we start reading part of a
line, it can no longer be sent out immediately, because it might be part
of a level prefix that needs to be stripped off. To solve this, the
parser code now does a *partial* buffering of all output: it will buffer
log data until it is certain that it is *not* part of a level prefix, at
which point all buffered data will be sent out to the otherwise
unbuffered outputs (terminal/stderr), even if a newline hasn't been
read yet.

steamrt/tasks#444

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>