test-logger: Demonstrate logging from a bash script
Unlike the current logger integration in `steam.sh`, this uses
`--background` to avoid conflicts with subreapers (steamrt/tasks#460),
and `--sh-syntax` to set environment variables (steamrt/tasks#461).

The unit test doesn't assert that the output includes the desired
environment variables, because we can't assume that the Journal is
always available in our test environments, and it's non-trivial to
mock up a terminal file descriptor - but when run on a system with a
working Journal, it does correctly set `SRT_LOGGER_USE_JOURNAL=1`.

steamrt/tasks#461

Signed-off-by: Simon McVittie <smcv@collabora.com>