logger: Add a mechanism to report when the logger is ready
When launching the logger in the background from a shell script, its
exit status will not always be available. There are three states we
need to signal to the caller: preparing, ready, and failed.

A caller can distinguish between preparing and ready/failed by polling
and reading standard output until EOF: until EOF is reached, the logger
is still preparing, and after EOF is reached, it is ready. But that
doesn't tell the caller whether startup was successful or not.

If the logger crashes before it has had a chance to send anything on
standard output, then standard output will be closed with no further
data; so we need to signal success by sending something non-empty.
Use shell syntax, so that we will be able to extend the format to
provide information to shell scripts such as `steam.sh`.

steamrt/tasks#460, steamrt/tasks#461

Signed-off-by: Simon McVittie <smcv@collabora.com>